Update all scripts to properly roll back changes
This commit makes a change to the `clean`, `generate-patch`, and `validate` due to a series of edge cases that were introduced from not properly rolling back changes from optional flags that were introduced in `prepare` during intermediary steps of each of these scripts.

To elaborate, the prepare script was normally structured in the following way:
1) Pull from upstream (or keep a local version)
2) Prepare subcharts
3) Apply patch
4) Optional: generateCRDChart changes
5) Optional: providesGVR changes

The changes from generateCRDChart and providesGVR need to be rolled back / rolled forward once complete on all three of these scripts (when cleaning, when generating a patch, and when checking if git is dirty during a validate), so all three of these scripts needed to be changed to consistently make this happen.