Create default roles for accessing Monitoring UIs
This commit adds the minimal permissions required for users to be able to access the various UIs for accessing Prometheus, Grafana, and Alertmanager via a kubectl proxy.

In order to access this via the Rancher UI, please see https://github.com/rancher/dashboard/issues/2408 for more details on what permissions are required.